make each config entry have their own informers
This way, informers can be namespaced, which should be lighter on the
apiserver.

As a tradeoff, one node informer will be created for each entry. This
can be prevented by having one shared, non-namespaced factory for nodes
and per-entry factories for pods, but I chose to not optimize
prematurely for now